notifier/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile setup.cfg
#
anyio==3.5.0
    # via httpcore
attrs==21.4.0
    # via notifier (setup.cfg)
certifi==2021.10.8
    # via
    #   httpcore
    #   httpx
charset-normalizer==2.0.12
    # via httpx
dnspython==2.2.1
    # via email-validator
email-validator==1.2.0
    # via pydantic
h11==0.12.0
    # via httpcore
httpcore==0.14.7
    # via httpx
httpx==0.22.0
    # via notifier (setup.cfg)
idna==3.3
    # via
    #   anyio
    #   email-validator
    #   rfc3986
pydantic[email]==1.9.0
    # via notifier (setup.cfg)
python-dotenv==0.20.0
    # via notifier (setup.cfg)
rfc3986[idna2008]==1.5.0
    # via httpx
schedule==1.1.0
    # via notifier (setup.cfg)
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
structlog==21.5.0
    # via notifier (setup.cfg)
typing-extensions==4.2.0
    # via pydantic
